requirements.txt
numpy==1.23.5
torch==2.0.0
torchtext==0.15.1
tqdm==4.64.1
jupyter
tokenizers~=0.13.3
transformers~=4.29.1
PyYAML~=6.0